+++ examples/user-admin-app/web/users/board/{name}/detail.sql
@@ -1,0 +1,9 @@
+select
+  u.name,
+  u.status,
+  u.created_at
+from
+  users u
+where
+  u.name = /* name */ 'sato'
+;