+++ examples/user-admin-app/web/users/board/{name}/groups.sql
@@ -1,0 +1,7 @@
+select
+  g.display_name
+from
+  app_groups g
+order by
+  g.display_name
+;